requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-redis==5.2.0
djangorestframework==3.13.1
httpie==2.6.0
idna==3.3
psycopg2==2.9.2
pycodestyle==2.8.0
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
redis==4.0.2
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.3
